pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "sigma"
version = "0.1.0"
description = "SIGMA: Self-supervised In-context Grounded Map Aggregation"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "opencv-python",
    "scikit-learn",
    "scikit-image",
    "hydra-core",
    "omegaconf",
    "pillow",
    "matplotlib",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sigma*"]
exclude = ["tests*", "docs*", "third_party*"]

[project.scripts]
sigma = "sigma.main:run_pipeline"
